% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ike_observation}
\alias{ike_observation}
\title{A Single Observation of Hurricane Ike}
\format{A data frame with columns:
\describe{
 \item{storm_id}{Storm Name and Year}
 \item{date}{Date.}
 \item{latitude}{Latitude.}
 \item{longitude}{Longitude.}
 \item{wind_speed}{Wind Speed}
 \item{ne}{Radius of wind speed.}
 \item{nw}{Radius of wind speed.}
 \item{se}{Radius of wind speed.}
 \item{sw}{Radius of wind speed.}
}}
\source{
Regional and Mesocale Meteorology Branch of Colorado
State University. \url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/}
}
\usage{
ike_observation
}
\description{
A Single Observation of Hurricane Ike
}
\examples{
\dontrun{
 ike_observation
}
}
\keyword{datasets}
