% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{geom_hurricane}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{other}
}
\description{
Generates a radii plot on a ggmap object representing the maximum windspeed
}
\examples{
\dontrun{
get_map("Louisiana", zoom = 6, maptype = "toner-background", source="stamen",force=FALSE) \%>\%
    ggmap(extent = "device") +
    geom_hurricane(data = katrina_observation,
                   aes(x = longitude, y = latitude,
                       r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
                       fill = wind_speed, color = wind_speed,
                       scale_radii=0.75)) +
    scale_color_manual(name = "Wind speed (kts)",
                       values = c("red", "orange", "yellow")) +
    scale_fill_manual(name = "Wind speed (kts)",
                      values = c("red", "orange", "yellow"))
}
}
